+++ src/validator/constraints/basic/not-blank.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { notBlank } from './not-blank';
-import { ConstraintContext } from '@/validator/constraint';
+import { ConstraintContext } from '../../constraint-validator';
 
 describe('notBlank', () => {
   it.each([
