+++ src/validator/constraints/basic/not-blank.ts
@@ -1,5 +1,5 @@
 import { Violation } from '@/validator/violation';
-import { ConstraintContext } from '@/validator/constraint';
+import { ConstraintContext } from '../../constraint-validator';
 
 const DEFAULT_MESSAGE = 'This value should not be blank.';
