+++ src/validator/constraints/other/compound.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from 'vitest';
 import { compound } from './compound';
 import { Violation } from '@/validator/violation';
-import { ConstraintContext } from '@/validator/constraint';
+import { ConstraintContext } from '../../constraint-validator';
 
 describe('compound', () => {
   it('creates a constraint that applies nested constraints from arrays', () => {
