+++ src/validator/constraints/other/compound.ts
@@ -1,5 +1,5 @@
 import { Violation } from '@/validator/violation';
-import { ConstraintContext } from '@/validator/constraint';
+import { ConstraintContext } from '../../constraint-validator';
 import { FieldSchema } from '@/validator/schema';
 
 export function compound(schema: FieldSchema) {
