+++ src/validator/constraints/string/slug.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { slug } from './slug';
-import { ConstraintContext } from '@/validator/constraint';
+import { ConstraintContext } from '../../constraint-validator';
 import { ConstraintOptions } from '@/validator/schema';
 
 describe('slug', () => {
