+++ src/validator/factory/create-validator.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'vitest';
 import { UnknownConstraintError } from '../errors';
 import { createValidator } from './create-validator';
 import { Violation } from '@/validator/violation';
-import { ConstraintContext } from '@/validator/constraint';
+import { ConstraintContext } from '../constraint-validator';
 
 describe('Validator', () => {
   describe('Registry', () => {
