+++ src/validator/index.ts
@@ -6,6 +6,6 @@ export { createValidationMiddleware } from './middleware/create-validation-middl
 
 export type { Validator } from './validator';
 export type { Payload } from './payload';
-export type { ConstraintValidator, ConstraintContext } from './constraint';
+export type { ConstraintValidator, ConstraintContext } from './constraint-validator';
 export type { ConstraintOptions, FieldSchema, ValidationSchema } from './schema';
 export type { Violation } from './violation';
